Don't ensure labels in testset instances
Since test (text) instances is read separately the order of the classes may be figgerent than in the training set.
But this does not matter since, the classes used in the evaluation is from the combined loading of train and test
set and the mapping used comes from the training set.